+++ source/Internal/Framework/Module/Configuration/DataObject/ModuleConfiguration.php
@@ -14,8 +14,6 @@ use OxidEsales\EshopCommunity\Internal\Framework\Module\Configuration\DataObject
 use OxidEsales\EshopCommunity\Internal\Framework\Module\Configuration\DataObject\ModuleConfiguration\Event;
 use OxidEsales\EshopCommunity\Internal\Framework\Module\Configuration\DataObject\ModuleConfiguration\SmartyPluginDirectory;
 use OxidEsales\EshopCommunity\Internal\Framework\Module\Configuration\DataObject\ModuleConfiguration\Template;
-use OxidEsales\EshopCommunity\Internal\Framework\Module\Configuration\DataObject\ModuleConfiguration\TemplateBlock;
-use OxidEsales\EshopCommunity\Internal\Framework\Module\Configuration\Exception\InvalidModuleIdException;
 use OxidEsales\EshopCommunity\Internal\Framework\Module\Configuration\Exception\ModuleSettingNotFountException;
 use OxidEsales\EshopCommunity\Internal\Framework\Module\Setting\Setting;
 use function Symfony\Component\String\u;
@@ -90,11 +88,6 @@ class ModuleConfiguration
      * @var SmartyPluginDirectory[]
      */
     private $smartyPluginDirectories = [];
-
-    /**
-     * @var TemplateBlock[]
-     */
-    private $templateBlocks = [];
 
     /**
      * @var Event[]
@@ -352,29 +345,6 @@ class ModuleConfiguration
     public function hasClassExtensions(): bool
     {
         return !empty($this->classExtensions);
-    }
-
-    /**
-     * @return TemplateBlock[]
-     */
-    public function getTemplateBlocks(): array
-    {
-        return $this->templateBlocks;
-    }
-
-    public function addTemplateBlock(TemplateBlock $templateBlock)
-    {
-        $this->templateBlocks[] = $templateBlock;
-
-        return $this;
-    }
-
-    /**
-     * @return bool
-     */
-    public function hasTemplateBlocks(): bool
-    {
-        return !empty($this->templateBlocks);
     }
 
     /**